// src/app/back-service/model/meetingDTO.ts
/**
 * Member Management Service API
 * MMS is application intended to register and manager members, record attendance and track contribution of each members. It also allow to schedule meetings and send notification to members of the church.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: metyh2@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface MeetingDTO { 
    address1?: string;
    address2?: string;
    agenda?: string;
    cellPhone?: string;
    city?: string;
    dateTime?: Date;
    duration?: number;
    email?: string;
    id?: number;
    state?: string;
    title?: string;
    zipCode?: string;
}